backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
gunicorn
certifi

# New SQL Libraries
sqlalchemy
psycopg2-binary
alembic

# FastAPI Users with SQLAlchemy support
fastapi-users[sqlalchemy]
fastapi-users-db-sqlalchemy